Fix: troca o tipo primitivo Long para Double

+++ src/main/java/com/bancomachado/bancomachado/controller/Response/SavingsAccountResponse.java
@@ -3,5 +3,5 @@ package com.bancomachado.bancomachado.controller.Response;
 import lombok.Builder;
 
 @Builder
-public record SavingsAccountResponse(Long id, int agency, int account, Long balance) {
+public record SavingsAccountResponse(Long id, int agency, int account, Double balance) {
 }
